1.11.1b
- Fixed toggle for off-hand usage (and a few bugs with it as well)
- Changed "ItemManager" to "CivItem" for some consistancy
- Updated some places to use new back head instead of back map
- Began major updating of materials from ids to Material (for 1.13)
- /res exchange now uses Material
- Removed TitleAPI dependencies
- Added Minecraft Titles to replace TitleAPI
- Enable eating of Golden Apple types (custom effects to them now! Still not craftable yet)
- Changed Blacksmith to use Material
- Cleaned Blacksmith smelting up
- Changed hunger and saturation values to some foods
- Simplified server MOTD on ping event
- Removed all internal CivData IDs

+++ civcraft/src/com/avrgaming/civcraft/structure/ScoutTower.java
@@ -72,7 +72,7 @@ public class ScoutTower extends Structure {
 			proximityComponent.setCenter(this.getCenterLocation());
 			proximityComponent.setRadius(range);
 			
-			reportSeconds = (int)CivSettings.getDouble(CivSettings.warConfig, "scout_tower.update");
+			reportSeconds = CivSettings.getInteger(CivSettings.warConfig, "scout_tower.update");
 			
 			
 		} catch (InvalidConfiguration e) {
